Walls
Can place walls by selecting tiles and then pressing w.
Walls set their cost to 200 so the enemies most likely won't path there.
They only will if there is no other way.

Tile
dir to next_pos
Instead of storing the dir to the next pos, store its position
Enemy keeps track of this position and only updates it when they get close
to it.
This makes the enemies move to the center of tiles instead of their edges.

Diagonal Walls
Might be bugged?
If 2 walls are connected at a diagonal, enemies can path through the
corner.
But they don't seem to really priorotize this. They only do it from the
nearest tiles. If they are a tile away, they will path all the way around.
That doesn't seem right.
I don't want them going through diagonal walls, but it being bugged might
mean other stuff is wrong.

move_enemy
Update your position if you are near your destination or you don't have
one. Otherwise, keep on your direction.
Set your direction when you get a new position to path to. This is what
moves you. Pos is there to figure out if you've reached your destination

generate_flow_field
removed the first version. This was the one copied from Minesweeper TD.

+++ src/enemy.rs
@@ -11,7 +11,21 @@ impl Plugin for EnemyPlugin {
 }
 
 #[derive(Component)]
-struct Enemy;
+struct Enemy {
+    dir: Vec2,
+    pos: Option<Vec2>,
+    speed: f32,
+}
+
+impl Enemy {
+    fn new() -> Self {
+        Enemy {
+            dir: Vec2::new(1.0, 0.0),
+            pos: None,
+            speed: 25.0,
+        }
+    }
+}
 
 fn spawn_enemy(mut commands: Commands, keyboard: Res<Input<KeyCode>>) {
     if keyboard.just_pressed(KeyCode::E) {
@@ -33,7 +47,7 @@ fn spawn_enemy(mut commands: Commands, keyboard: Res<Input<KeyCode>>) {
                         transform: Transform::from_translation(offset + pos),
                         ..default()
                     })
-                    .insert(Enemy);
+                    .insert(Enemy::new());
             }
         }
     }
@@ -42,21 +56,45 @@ fn spawn_enemy(mut commands: Commands, keyboard: Res<Input<KeyCode>>) {
 fn move_enemy(
     grid: Res<Grid>,
     q_tiles: Query<(&Transform, &Tile), Without<Enemy>>,
-    mut q_enemies: Query<&mut Transform, With<Enemy>>,
+    mut q_enemies: Query<(&mut Transform, &mut Enemy)>,
     time: Res<Time>,
 ) {
-    for mut trans in q_enemies.iter_mut() {
-        let mut dir = Vec3::new(-25.0, 0.0, 0.0);
+    for (mut trans, mut enemy) in q_enemies.iter_mut() {
+        let mut want_pos = false;
+        if let Some(pos) = enemy.pos {
+            // you have a pos
+            if trans.translation.truncate().distance_squared(pos) < 1.0 {
+                // close to pos
+                // find a new one
+                want_pos = true;
+            } else {
+                // keep going where you're going
+            }
+        } else {
+            // find one if possible
+            want_pos = true;
+        }
 
-        if let Some(entity) = grid.get_vec2(trans.translation.truncate()) {
-            if let Ok((_transform, tile)) = q_tiles.get(entity) {
-                if let Some(dir_tile) = tile.dir {
-                    // trans.translation += dir.extend(0.0) * time.delta_seconds();
-                    dir = dir_tile.extend(0.0) * 25.0;
+        // let mut dir = Vec3::new(-25.0, 0.0, 0.0);
+        if want_pos {
+            if let Some(entity) = grid.get_vec2(trans.translation.truncate()) {
+                if let Ok((_transform, tile)) = q_tiles.get(entity) {
+                    if let Some(dir_tile) = tile.next_pos {
+                        // trans.translation += dir.extend(0.0) * time.delta_seconds();
+                        // dir = dir_tile.extend(0.0) * 25.0;
+                        enemy.pos = Some(dir_tile);
+                        enemy.dir = (dir_tile - trans.translation.truncate()).normalize_or_zero();
+                    } else {
+                        // at destination
+                        // stop moving
+                        if enemy.dir != Vec2::ZERO {
+                            enemy.dir = Vec2::ZERO;
+                        }
+                    }
                 }
             }
         }
 
-        trans.translation += dir * time.delta_seconds();
+        trans.translation += enemy.dir.extend(0.0) * enemy.speed * time.delta_seconds();
     }
 }
